Rename pinfo to prte_info to remove conflicts
