+++ orte/tools/prte_info/help-orte-info.txt
@@ -11,7 +11,7 @@
 # Copyright (c) 2004-2005 The Regents of the University of California.
 #                         All rights reserved.
 # Copyright (c) 2010-2012 Cisco Systems, Inc.  All rights reserved.
-# Copyright (c) 2016      Intel, Inc.  All rights reserved.
+# Copyright (c) 2016-2018 Intel, Inc. All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
