+++ orte/tools/prte_info/output.c
@@ -10,6 +10,7 @@
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
  * Copyright (c) 2010      Cisco Systems, Inc.  All rights reserved.
+ * Copyright (c) 2018      Intel, Inc. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -33,7 +34,7 @@
 #endif
 #include <ctype.h>
 
-#include "orte/tools/pinfo/pinfo.h"
+#include "orte/tools/prte_info/pinfo.h"
 
 #include "opal/util/show_help.h"
