+++ orte/tools/prte_info/pinfo.1
@@ -1,0 +1,162 @@
+.\" Man page contributed by Dirk Eddelbuettel <edd@debian.org>
+.\" and released under the BSD license
+.\" Copyright (c) 2008      Sun Microsystems, Inc.  All rights reserved.
+.TH orte-info 1 "@OMPI_RELEASE_DATE@" "gitclone" "PMIx-Runtime-Environment"
+.SH NAME
+orte-info - Display information about the ORTE installation
+.
+.\" **************************
+.\"    Synopsis Section
+.\" **************************
+.SH SYNOPSIS
+.B orte-info [options]
+.
+.\" **************************
+.\"    Description Section
+.\" **************************
+.SH DESCRIPTION
+.PP
+.B orte-info
+provides detailed information about the ORTE installation.  It can
+be useful for at least three common scenarios:
+.PP
+1. Checking local configuration and seeing how ORTE was installed.
+.PP
+2. Submitting bug reports / help requests to the ORTE community
+(see
+.IR http://www.open-mpi.org/community/help/ )
+.PP
+3. Seeing a list of installed ORTE plugins and querying what
+MCA parameters they support.
+.
+.\" **************************
+.\"    Options Section
+.\" **************************
+.SH OPTIONS
+.B orte-info
+accepts the following options:
+.TP 8
+.I \-a|\-\-all
+Show all configuration options and MCA parameters
+.TP 8
+.I \-\-arch
+Show architecture ORTE was compiled on
+.TP 8
+.I \-c|\-\-config
+Show configuration options
+.TP 8
+.I \-gmca|\-\-gmca <param> <value>
+Pass global MCA parameters that are applicable to all contexts.
+.TP 8
+.I \-h|\-\-help
+Shows help / usage message
+.TP 8
+.I \-\-hostname
+Show the hostname that ORTE was configured and built on
+.TP 8
+.I \-\-internal
+Show internal MCA parameters (not meant to be modified by users)
+.TP 8
+.I \-mca|\-\-mca <param> <value>
+Pass context-specific MCA parameters; they are considered global if --gmca is
+not used and only one context is specified.
+.TP 8
+.I \-\-param <type> <component>
+Show MCA parameters.  The first parameter is the type of the component
+to display; the second parameter is the specific component to display
+(or the keyword "all", meaning "display all components of this type").
+.TP 8
+.I \-\-parsable
+When used in conjunction with other parameters, the output is
+displayed in a machine-parsable format
+.I \-\-parseable
+Synonym for --parsable
+.TP 8
+.I \-\-path <type>
+Show paths that ORTE was configured with.  Accepts the following
+parameters: prefix, bindir, libdir, incdir, pkglibdir, sysconfdir.
+.TP 8
+.I \-\-pretty
+When used in conjunction with other parameters, the output is
+displayed in 'prettyprint' format (default)
+.TP 8
+.I \-v|\-\-version <component> <scope>
+Show version of ORTE or a component.  <component> can be the
+keywords "ompi" or "all", the name of a framework (e.g., "coll" shows
+all components in the coll framework), or the name of a specific
+component (e.g., "pls:rsh" shows the information from the rsh PLS
+component).  <scope> can be one of: full, major, minor, release,
+greek, svn.
+.
+.\" **************************
+.\"    Examples Section
+.\" **************************
+.SH EXAMPLES
+.TP 4
+orte-info
+Show the default output of options and listing of installed
+components in a human-readable / prettyprint format.
+.
+.
+.TP
+orte-info --parsable
+Show the default output of options and listing of installed
+components in a machine-parsable format.
+.
+.
+.TP
+orte-info --param rmcast udp
+Show the MCA parameters of the "udp" RMCAST component in a
+human-readable / prettyprint format.
+.
+.
+.TP
+orte-info --param rmcast udp --parsable
+Show the MCA parameters of the "udp" RMCAST component in a
+machine-parsable format.
+.
+.
+.TP
+orte-info --path bindir
+Show the "bindir" that ORTE was configured with.
+.
+.
+.TP
+orte-info --version orte full --parsable
+Show the full version numbers of ORTE (including the OPAL
+version number) in a machine-readable format.
+.
+.
+.TP
+orte-info --version rmcast major
+Show the major version number of all RMCAST components in a prettyprint
+format.
+.
+.
+.TP
+orte-info --version rmcast:tcp minor
+Show the minor version number of the TCP RMCAST component in a
+prettyprint format.
+.
+.
+.TP
+orte-info --all
+Show
+.I all
+information about the ORTE installation, including all components
+that can be found, the MCA parameters that they support, versions of
+ORTE and the components, etc.
+.
+.
+.\" **************************
+.\"    Authors Section
+.\" **************************
+.SH AUTHORS
+The ORTE maintainers -- see
+.I http://www.openmpi.org/
+or the file
+.IR AUTHORS .
+.PP
+This manual page was originally contributed by Dirk Eddelbuettel
+<edd@debian.org>, one of the Debian GNU/Linux maintainers for Open
+MPI, and may be used by others.
